fix: betty
